+++ angular-view/src/app/app-routing.module.ts
@@ -4,7 +4,10 @@ import { RouterModule, Routes } from '@angular/router';
 import { MainComponent } from './components/main/main.component';
 
 const routes: Routes = [
-  { path: '', component: MainComponent }
+  { path: 'index', component: MainComponent },
+  { path: '', component: MainComponent },
+  //{ path: '**', component: PageNotFoundComponent }
+  // TODO
 ];
 
 @NgModule({