+++ express-server/auth/config.jwt.js
@@ -19,7 +19,7 @@ const options = {
 // The JWT payload is passed into the verify callback
 passport.use(new JwtStrategy(options, function (jwt_payload, done) {
 
-  console.log(jwt_payload);
+  //console.log(jwt_payload);
 
   // We will assign the `sub` property on the JWT to the database ID of user
   User.findOne({ _id: jwt_payload.sub }, function (err, user) {
@@ -34,7 +34,6 @@ passport.use(new JwtStrategy(options, function (jwt_payload, done) {
     } else {
       return done(null, false);
     }
-
   });
 
 }));