+++ express-server/routes/auth.js
@@ -1,12 +1,24 @@
 const router = require('express').Router();
 const passport = require('passport');
 const querystring = require("querystring");
+const User = require('mongoose').model('User');
 
 const CONFIG = require('../config');
 const utils = require('../lib/utils');
 
-router.get('/protected', passport.authenticate('jwt', { session: false }), (req, res, next) => {
-  res.status(200).json({ success: true, msg: "You are successfully authenticated to this route!" });
+router.get('/protected', passport.authenticate('jwt', { session: false }), async (req, res, next) => {
+  console.log("--protected");
+
+  const user = await User.findOne({ _id: req.user._id })
+    .exec()
+    .then((user) => {
+      return user;
+    })
+    .catch((err) => {
+      return 'error occured';
+    });
+
+  res.status(200).json(user);
 });
 router.get('/test', (req, res, next) => {
   res.status(200).json({ success: true, msg: "You are successfully authenticated to this route!" });
@@ -25,11 +37,9 @@ router.get('/' + CONFIG.oAuth_redirect_postfix,
     const tokenObject = utils.issueJWT(req.user);
     const data = { success: true, token: tokenObject.token, expiresIn: tokenObject.expires };
 
-    console.log("test");
     const escape = encodeURIComponent(JSON.stringify(data));
-    console.log(escape);
 
-    res.redirect('/t?token=' + escape);
+    res.redirect('/?authtoken=' + escape);
   });
 
 module.exports = router;